Add D3D11 support for textures and uniform buffers
- Enable D3D11UniformBuffer creation in UniformBuffer factory
- Add TextureUtils for Kerberos-to-DXGI format mapping
- Use correct DXGI formats for D3D11 textures (not hardcoded)
- Set debug names for D3D11 textures and SRVs via SetPrivateData
- Switch font atlas to RGBA8 and 4-channel MSDF generator
- Cast data size in OpenGL texture upload for type safety
- Minor cleanups and function signature improvements

+++ Kerberos/src/Platform/OpenGL/OpenGLTexture.cpp
@@ -91,7 +91,7 @@ namespace Kerberos
 		if (data)
 		{
 			/// Upload the texture data to the GPU
-			OpenGLTexture2D::SetData(data.Data, data.Size);
+			OpenGLTexture2D::SetData(data.Data, static_cast<uint32_t>(data.Size));
 		}
 	}
